feat: ssp 계산 공식 기능 적용 - 수정필요

+++ src/main/java/com/nsmm/esg/tcfdservice/dto/StrategyScenarioAnalysisRequest.java
@@ -19,7 +19,7 @@ public class StrategyScenarioAnalysisRequest {
     private final Integer baseYear;
     private final String climate;
     private final Double assetValue;
-    private final Double estimatedDamage;
+    private final Long estimatedDamage;
 
 
     public StrategyScenarioAnalysis toEntity(Long memberId) {
